engine: split db/tests.rs to restore per-file line-cap headroom
db/tests.rs had grown to 5995 lines, five short of the 6000-line
per-file cap that scripts/lint-project.sh enforces in pre-commit and
CI, so the next non-trivial test addition would have blocked an
unrelated commit with a forced split under duress.

Move two cohesive sections (along the existing section headers) into
sibling test modules, following the db/ precedent of one file per test
concern: the incremental-compilation + AC acceptance-criteria tests
into incremental_compile_tests.rs, and the PREVIOUS/INIT opcode + LTM
incremental-verification tests into vm_verification_tests.rs. The move
is behavior-preserving: test bodies are byte-identical, only the
simple_project helper gains pub(crate) visibility so the siblings can
keep sharing it. tests.rs lands at 2943 lines.

Fixes #645
