+++ src/simlin-engine/src/db.rs
@@ -983,6 +983,8 @@ mod dimension_invalidation_tests;
 #[cfg(test)]
 mod fragment_cache_tests;
 #[cfg(test)]
+mod incremental_compile_tests;
+#[cfg(test)]
 mod ltm_module_tests;
 #[cfg(test)]
 mod ltm_unified_tests;
@@ -990,3 +992,5 @@ mod ltm_unified_tests;
 mod prev_init_tests;
 #[cfg(test)]
 mod tests;
+#[cfg(test)]
+mod vm_verification_tests;
